mi_iommu: fix Clang constant conversion error
- Cast MAX_ALIGN(iova_shift(iovad)) to unsigned int when
  passing to iova_get_pad_size(), to silence Clang warning
  about implicit conversion from unsigned long (ULONG_MAX)
  to unsigned int.

This resolves Clang 18 build error in mi_iommu.c while
preserving intended alignment logic.

+++ drivers/iommu/mi_iommu.c
@@ -187,7 +187,7 @@ struct rb_node *rdxtree_matched_gap(struct iova_domain *iovad,
 				}
 
 				if (size_aligned)
-					pad_size = iova_get_pad_size(size, limit_size, MAX_ALIGN(iova_shift(iovad)));
+					pad_size = iova_get_pad_size(size, limit_size,(unsigned int)MAX_ALIGN(iova_shift(iovad)));
 
 				/*
 				 *  best fit
